refactor(auth): move getPastTrades method
