+++ lib/authenticated.js
@@ -53,29 +53,6 @@ class AuthenticatedClient extends PublicClient {
   }
 
   /**
-   * @param {Object} [options={}]
-   * @param {string} [options.symbol] - The symbol to retrieve trades for.
-   * @param {number} [options.limit_trades] - The maximum number of trades to return.
-   * @param {number} [options.timestamp] - Only return trades on or after this timestamp.
-   * @example
-   * const trades = await authClient.getPastTrades();
-   * @description Get your past trades.
-   * @see {@link https://docs.gemini.com/rest-api/#get-past-trades|get-past-trades}
-   */
-  getPastTrades({
-    symbol = this.symbol,
-    limit_trades = API_LIMIT,
-    timestamp
-  } = {}) {
-    return this.post({
-      request: "/v1/mytrades",
-      symbol,
-      limit_trades,
-      timestamp
-    });
-  }
-
-  /**
    * @example
    * const volume = await authClient.getNotionalVolume();
    * @description Get the volume in price currency that has been traded across all pairs over a period of 30 days.
